Update by Chinese Ghost Day (Zhongyuan Festival)
Mourn to those we have lost! We have something to defend! We have a reason to live!
Define NoirVisor's UEFI Variable Vendor GUID.
Optimize the assembly code part for VM-Exit Handlers in VT-Core and SVM-Core.
Fix bug that EPT is not constructed properly if stealth hook with EPT is disabled.
Fix bug so that the correct CR3 is switched during hypercall.
Add a handler in case vmresume instruction fails in VT-Core.
Remove static IAT dependency to the HalGetEnvironmentVariableEx function.
Add UMIP Emulation (aka. NPIEP) development plan in readme.