Update by Mar.3rd, 2022
Porting to UEFI. (Incomplete due to an unknown bug that, after subversion, shuts down the virtual machine without an error notice.)
Fixed intrinsics that LLVM did not implemented.
Rewrote the subversion routine in assembly to circumvent the bug in LLVM.
Moved the MSR-Hook Assembly code to Windows folder because it is not universal.
Implemented host function wrappers in UEFI.
Implemented serial I/O.